Refactor: corosync: move conf_parser.ConfParser to corosync.ConfParser
