+++ crmsh/corosync.py
@@ -8,12 +8,12 @@ configuration file, and also the corosync-* utilities.
 import os
 import re
 import typing
+from io import StringIO
 
 from . import utils, sh
 from . import tmpfiles
 from . import parallax
 from . import log
-from . import conf_parser
 from . import corosync_config_format
 from .sh import ShellUtils
 
@@ -22,6 +22,22 @@ logger = log.setup_logger(__name__)
 
 
 COROSYNC_TOKEN_DEFAULT = 1000  # in ms units
+COROSYNC_CONF_TEMPLATE = """
+totem {
+    version: 2
+}
+
+quorum {
+    provider: corosync_votequorum
+}
+
+logging {
+    to_logfile: yes
+    logfile: /var/log/cluster/corosync.log
+    to_syslog: yes
+    timestamp: on
+}
+"""
 
 
 def is_knet() -> bool:
@@ -36,7 +52,7 @@ def is_using_ipv6() -> bool:
 
 def get_link_number() -> int:
     link_num = 1
-    for key, value in conf_parser.ConfParser.get_value("nodelist.node").items():
+    for key, value in ConfParser.get_value("nodelist.node").items():
         if re.search("ring[1-7]_addr", key) and value:
             link_num += 1
     return link_num
@@ -217,15 +233,15 @@ def get_free_nodeid():
 
 
 def get_value(path, index: int = 0):
-    return conf_parser.ConfParser.get_value(path, index)
+    return ConfParser.get_value(path, index)
 
 
 def get_values(path):
-    return conf_parser.ConfParser.get_values(path)
+    return ConfParser.get_values(path)
 
 
 def set_value(path, value, index: int = 0):
-    conf_parser.ConfParser.set_value(path, value, index)
+    ConfParser.set_value(path, value, index)
 
 
 class IPAlreadyConfiguredError(Exception):
@@ -262,7 +278,7 @@ def add_node_config(ip_list: typing.List[str]) -> None:
     Add nodelist in corosync.conf
     """
     find_configured_ip(ip_list)
-    inst = conf_parser.ConfParser()
+    inst = ConfParser()
     node_index = len(inst.get_all("nodelist.node"))
     for i, addr in enumerate(ip_list):
         inst.set("nodelist.node.ring{}_addr".format(i), addr, node_index)
@@ -275,7 +291,7 @@ def del_node(addr: str) -> None:
     '''
     Remove node from corosync
     '''
-    inst = conf_parser.ConfParser()
+    inst = ConfParser()
     name_list = inst.get_all("nodelist.node.ring0_addr")
     index = name_list.index(addr)
     inst.remove("nodelist.node", index)
@@ -319,6 +335,154 @@ def token_and_consensus_timeout():
 
 
 def get_all_paths():
-    with open(conf(), 'r', encoding='utf-8') as f:
-        dom = corosync_config_format.DomParser(f).dom()
-        return corosync_config_format.DomQuery(dom).enumerate_all_paths()
+    return ConfParser().dom_query().enumerate_all_paths()
+
+
+class ConfParser(object):
+    """
+    Class to parse config file which format like corosync.conf
+    """
+    COROSYNC_KNOWN_SEC_NAMES_WITH_LIST = {("totem", "interface"), ("nodelist", "node")}
+
+    def __init__(self, config_file=None, config_data=None, sec_names_with_list=()):
+        self._config_file = config_file
+        self._sec_names_with_list = set(sec_names_with_list) if sec_names_with_list else self.COROSYNC_KNOWN_SEC_NAMES_WITH_LIST
+        if config_data is not None:
+            self._dom = corosync_config_format.DomParser(StringIO(config_data)).dom()
+        else:
+            if config_file:
+                self._config_file = config_file
+            else:
+                self._config_file = conf()
+            try:
+                with open(self._config_file, 'r', encoding='utf-8') as f:
+                    self._dom = corosync_config_format.DomParser(f).dom()
+            except OSError as e:
+                raise ValueError(str(e)) from None
+
+        self._dom_query = corosync_config_format.DomQuery(self._dom)
+
+    def dom_query(self):
+        return self._dom_query
+
+    def save(self, config_file=None, file_mode=0o644):
+        """save the config to config file"""
+        if not config_file:
+            config_file = self._config_file
+        with utils.open_atomic(config_file, 'w', fsync=True, encoding='utf-8') as f:
+            corosync_config_format.DomSerializer(self._dom, f)
+            os.fchmod(f.fileno(), file_mode)
+
+    def get(self, path, index=0):
+        """
+        Gets the value for the path
+
+        path: config path
+        index: known index in section
+        """
+        try:
+            return self._dom_query.get(path, index)
+        except (KeyError, IndexError):
+            return None
+
+    def get_all(self, path):
+        """
+        Returns all values matching path
+        """
+        return self._dom_query.get_all(path)
+
+    def remove(self, path, index=0):
+        try:
+            self._dom_query.remove(path, index)
+        except (KeyError, IndexError):
+            raise ValueError("Cannot find value on path \"{}:{}\"".format(path, index)) from None
+
+    def _raw_set(self, path, value, index):
+        path = path.split('.')
+        node = self._dom
+        path_stack = tuple()
+        for key in path[:-1]:
+            path_stack = (*path_stack, key)
+            if key not in node:
+                new_node = dict()
+                node[key] = new_node
+                node = new_node
+            else:
+                match node[key]:
+                    case dict(_) as next_node:
+                        if index > 0 and path_stack in self._sec_names_with_list:
+                            if index == 1:
+                                new_node = dict()
+                                node[key] = [next_node, new_node]
+                                node = new_node
+                            else:
+                                raise IndexError(f'index out of range: {index}')
+                        else:
+                            node = next_node
+                    case list(_) as li:
+                        if index > len(li):
+                            raise IndexError(f'index out of range: {index}')
+                        elif index == len(li):
+                            new_node = dict()
+                            li.append(new_node)
+                            node = new_node
+                        else:
+                            node = li[index]
+        key = path[-1]
+        if key not in node:
+            node[key] = value
+        else:
+            match node[key]:
+                case list(_) as li:
+                    if index > len(li):
+                        raise IndexError(f'index out of range: {index}')
+                    elif index == len(li):
+                        li.append(value)
+                    else:
+                        li[index] = value
+                case _:
+                    node[key] = value
+
+    def set(self, path, value, index=0):
+        try:
+            self._raw_set(path, value, index)
+        except KeyError:
+            raise ValueError("Invalid path \"{}\"".format(path)) from None
+        except IndexError:
+            raise ValueError(f'Index {index} out of range at path "{path}"') from None
+
+    @classmethod
+    def get_value(cls, path: str, index: int = 0):
+        """
+        Class method to get value
+        Return None if not found
+        """
+        inst = cls()
+        return inst.get(path, index)
+
+    @classmethod
+    def get_values(cls, path: str):
+        """
+        Class method to get value list matched by path
+        Return [] if not matched
+        """
+        inst = cls()
+        return inst.get_all(path)
+
+    @classmethod
+    def set_value(cls, path, value, index=0):
+        """
+        Class method to set value for path
+        Then write back to config file
+        """
+        inst = cls()
+        inst.set(path, value, index)
+        inst.save()
+
+    @classmethod
+    def remove_key(cls, path, index=0):
+        """
+        """
+        inst = cls()
+        inst.remove(path, index)
+        inst.save()
