+++ data-manifest
@@ -188,7 +188,6 @@ test/unittests/test_bugs.py
 test/unittests/test_cib.py
 test/unittests/test_cliformat.py
 test/unittests/test.conf
-test/unittests/test_conf_parser.py
 test/unittests/test_corosync_config_format.py
 test/unittests/test_corosync.py
 test/unittests/test_crashtest_check.py
